+++ core/src/main/java/com/graphhopper/routing/weighting/ShortestWeighting.java
@@ -21,34 +21,59 @@ import com.graphhopper.routing.ev.BooleanEncodedValue;
 import com.graphhopper.routing.ev.DecimalEncodedValue;
 import com.graphhopper.util.EdgeIteratorState;
 
-import static com.graphhopper.routing.weighting.TurnCostProvider.NO_TURN_COST_PROVIDER;
-
 /**
  * Calculates the shortest route - independent of a vehicle as the calculation is based on the
  * distance only.
- * <p>
  *
  * @author Peter Karich
  */
-public class ShortestWeighting extends AbstractWeighting {
+public class ShortestWeighting implements Weighting {
+
+    final BooleanEncodedValue accessEnc;
+    final DecimalEncodedValue speedEnc;
+    final TurnCostProvider tcProvider;
+
     public ShortestWeighting(BooleanEncodedValue accessEnc, DecimalEncodedValue speedEnc) {
-        this(accessEnc, speedEnc, NO_TURN_COST_PROVIDER);
+        this(accessEnc, speedEnc, TurnCostProvider.NO_TURN_COST_PROVIDER);
     }
 
-    public ShortestWeighting(BooleanEncodedValue accessEnc, DecimalEncodedValue speedEnc, TurnCostProvider turnCostProvider) {
-        super(accessEnc, speedEnc, turnCostProvider);
+    public ShortestWeighting(BooleanEncodedValue accessEnc, DecimalEncodedValue speedEnc, TurnCostProvider tcProvider) {
+        this.accessEnc = accessEnc;
+        this.speedEnc = speedEnc;
+        this.tcProvider = tcProvider;
     }
 
     @Override
-    public double getMinWeight(double currDistToGoal) {
-        return currDistToGoal;
+    public double getMinWeight(double distance) {
+        return distance;
     }
 
     @Override
     public double calcEdgeWeight(EdgeIteratorState edgeState, boolean reverse) {
-        if (edgeHasNoAccess(edgeState, reverse))
+        if (reverse ? !edgeState.getReverse(accessEnc) : !edgeState.get(accessEnc))
             return Double.POSITIVE_INFINITY;
         return edgeState.getDistance();
+    }
+
+    @Override
+    public long calcEdgeMillis(EdgeIteratorState edgeState, boolean reverse) {
+        double speed = reverse ? edgeState.getReverse(speedEnc) : edgeState.get(speedEnc);
+        return Math.round(edgeState.getDistance() / speed * 3.6 * 1000);
+    }
+
+    @Override
+    public double calcTurnWeight(int inEdge, int viaNode, int outEdge) {
+        return tcProvider.calcTurnWeight(inEdge, viaNode, outEdge);
+    }
+
+    @Override
+    public long calcTurnMillis(int inEdge, int viaNode, int outEdge) {
+        return tcProvider.calcTurnMillis(inEdge, viaNode, outEdge);
+    }
+
+    @Override
+    public boolean hasTurnCosts() {
+        return tcProvider != TurnCostProvider.NO_TURN_COST_PROVIDER;
     }
 
     @Override